Tests/develop

+++ src/QuantumBackend.cpp
@@ -49,7 +49,7 @@ std::vector<sys_sage::QuantumGate*> sys_sage::QuantumBackend::GetAllGateTypes() 
     return gate_types;
 }
 
-int sys_sage::QuantumBackend::GetNumberofGates() const { return gate_types.size(); }
+size_t sys_sage::QuantumBackend::GetNumberofGates() const { return gate_types.size(); }
 
 std::vector<sys_sage::Qubit *> sys_sage::QuantumBackend::GetAllQubits()
 {
